Move internal packages to internal/ dir
Fixes #57

Cleans up the top-level dir enough for now
