+++ dry.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 
 	jsonpatch "github.com/evanphx/json-patch/v5"
-	"github.com/manifestival/manifestival/patch"
+	"github.com/manifestival/manifestival/internal/patch"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
