+++ internal/sources/path_test.go
@@ -5,7 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	. "github.com/manifestival/manifestival/sources"
+	. "github.com/manifestival/manifestival/internal/sources"
 )
 
 func TestParsing(t *testing.T) {
@@ -17,38 +17,38 @@ func TestParsing(t *testing.T) {
 		wantError bool
 	}{{
 		name: "single directory",
-		path: "../testdata/tree",
+		path: "testdata/tree",
 		want: []string{"a", "b"},
 	}, {
 		name:      "single directory, recursive",
-		path:      filepath.FromSlash("../testdata/tree"),
+		path:      filepath.FromSlash("testdata/tree"),
 		recursive: true,
 		want:      []string{"foo", "bar", "baz", "a", "b"},
 	}, {
 		name:      "single file",
-		path:      filepath.FromSlash("../testdata/tree/dir/b.yaml"),
+		path:      filepath.FromSlash("testdata/tree/dir/b.yaml"),
 		recursive: true,
 		want:      []string{"bar", "baz"},
 	}, {
 		name:      "single file, recursive",
-		path:      filepath.FromSlash("../testdata/tree/file.yaml"),
+		path:      filepath.FromSlash("testdata/tree/file.yaml"),
 		recursive: true,
 		want:      []string{"a", "b"},
 	}, {
 		name:      "missing file",
-		path:      filepath.FromSlash("../testdata/missing"),
+		path:      filepath.FromSlash("testdata/missing"),
 		wantError: true,
 	}, {
 		name:      "dangling symlink",
-		path:      filepath.FromSlash("../testdata/dangling-symlink"),
+		path:      filepath.FromSlash("testdata/dangling-symlink"),
 		wantError: true,
 	}, {
 		name:      "dangling symlink",
-		path:      filepath.FromSlash("../testdata/dangling-symlink"),
+		path:      filepath.FromSlash("testdata/dangling-symlink"),
 		wantError: true,
 	}, {
 		name:      "absolute path",
-		path:      filepath.Join(os.Getenv("PWD"), filepath.FromSlash("../testdata/tree/dir/b.yaml")),
+		path:      filepath.Join(os.Getenv("PWD"), filepath.FromSlash("testdata/tree/dir/b.yaml")),
 		recursive: true,
 		want:      []string{"bar", "baz"},
 	}, {
@@ -65,16 +65,16 @@ func TestParsing(t *testing.T) {
 		want: []string{"a", "b", "foo"},
 	}, {
 		name: "url and file",
-		path: "https://raw.githubusercontent.com/manifestival/manifestival/master/testdata/tree/file.yaml,../testdata/tree/dir/a.yaml",
+		path: "https://raw.githubusercontent.com/manifestival/manifestival/master/testdata/tree/file.yaml,testdata/tree/dir/a.yaml",
 		want: []string{"a", "b", "foo"},
 	}, {
 		name:      "url and directory, recursive",
-		path:      "https://raw.githubusercontent.com/manifestival/manifestival/master/testdata/tree/file.yaml,../testdata/tree/dir",
+		path:      "https://raw.githubusercontent.com/manifestival/manifestival/master/testdata/tree/file.yaml,testdata/tree/dir",
 		recursive: true,
 		want:      []string{"a", "b", "foo", "bar", "baz"},
 	}, {
 		name:      "file and directory, recursive",
-		path:      filepath.FromSlash("../testdata/tree/file.yaml") + "," + filepath.FromSlash("../testdata/tree/dir"),
+		path:      filepath.FromSlash("testdata/tree/file.yaml") + "," + filepath.FromSlash("testdata/tree/dir"),
 		recursive: true,
 		want:      []string{"a", "b", "foo", "bar", "baz"},
 	}, {
