+++ internal/sources/testdata/dangling-symlink/dangling.yaml
