+++ internal/sources/testdata/dangling-symlink/valid.yaml
